% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldeRstructure.R
\name{foldeRstructure}
\alias{foldeRstructure}
\title{Create internal folder structure}
\usage{
foldeRstructure(Functions = TRUE, Rmd = TRUE, html = TRUE, pdf = TRUE)
}
\arguments{
\item{Functions}{Bolean Whether to create a subfolder Functions within Script folder. TRUE by default}

\item{Rmd}{Bolean Whether to create a subfolder Rmd within Output folder. TRUE by default}

\item{html}{Bolean Whether to create a subfolder html within Output folder. TRUE by default}

\item{pdf}{Bolean Whether to create a subfolder pdf within Output folder. TRUE by default}
}
\description{
This function creates an internal structure of working directory for RStudio projects following Martin Chan´s 
suggestion https://martinctc.github.io/blog/rstudio-projects-and-working-directories-a-beginner's-guide/
}
